telemetry: split histogram gtests into their own module
Move the Prometheus histogram tests (functional observation, bucket env
override, and the bucket-parsing unit tests) out of telemetry_prometheus_test.cpp
into a new telemetry_histogram_test.cpp, and extract the shared exporter fixture
into prometheus_telemetry_fixture.h so both files drive the same gtest suite
(SetUpTestSuite runs once, so the plugin directory is registered a single time).

Improves clarity by isolating the histogram tests, and keeps the PR under the
line cap since the histogram tests now live in new files. The relocated tests
scrape via the shared timeseries parser instead of the prometheus file's private
string helpers.
